Save 64-bits of data for each applet instead of 32

+++ software/o_c_REV/HEM_Calculate.ino
@@ -92,14 +92,14 @@ public:
         rand_clocked[selected] = 0;
     }
 
-    uint32_t OnDataRequest() {
-        uint32_t data = 0;
+    uint64_t OnDataRequest() {
+        uint64_t data = 0;
         Pack(data, PackLocation {0, 8}, operation[0]);
         Pack(data, PackLocation {8, 8}, operation[1]);
         return data;
     }
 
-    void OnDataReceive(uint32_t data) {
+    void OnDataReceive(uint64_t data) {
         operation[0] = Unpack(data, PackLocation {0, 8});
         operation[1] = Unpack(data, PackLocation {8, 8});
     }
@@ -169,10 +169,10 @@ void Calculate_ToggleHelpScreen(bool hemisphere) {
     Calculate_instance[hemisphere].HelpScreen();
 }
 
-uint32_t Calculate_OnDataRequest(bool hemisphere) {
+uint64_t Calculate_OnDataRequest(bool hemisphere) {
     return Calculate_instance[hemisphere].OnDataRequest();
 }
 
-void Calculate_OnDataReceive(bool hemisphere, uint32_t data) {
+void Calculate_OnDataReceive(bool hemisphere, uint64_t data) {
     Calculate_instance[hemisphere].OnDataReceive(data);
 }
