+++ software/o_c_REV/HEM_VectorEG.ino
@@ -81,15 +81,15 @@ public:
         }
     }
         
-    uint32_t OnDataRequest() {
-        uint32_t data = 0;
+    uint64_t OnDataRequest() {
+        uint64_t data = 0;
         Pack(data, PackLocation {0,6}, waveform_number[0]);
         Pack(data, PackLocation {6,6}, waveform_number[1]);
         Pack(data, PackLocation {12,10}, freq[0] & 0x03ff);
         Pack(data, PackLocation {22,10}, freq[1] & 0x03ff);
         return data;
     }
-    void OnDataReceive(uint32_t data) {
+    void OnDataReceive(uint64_t data) {
         freq[0] = Unpack(data, PackLocation {12,10});
         freq[1] = Unpack(data, PackLocation {22,10});
         SwitchWaveform(0, Unpack(data, PackLocation {0,6}));
@@ -195,5 +195,5 @@ void VectorEG_View(bool hemisphere) {VectorEG_instance[hemisphere].BaseView();}
 void VectorEG_OnButtonPress(bool hemisphere) {VectorEG_instance[hemisphere].OnButtonPress();}
 void VectorEG_OnEncoderMove(bool hemisphere, int direction) {VectorEG_instance[hemisphere].OnEncoderMove(direction);}
 void VectorEG_ToggleHelpScreen(bool hemisphere) {VectorEG_instance[hemisphere].HelpScreen();}
-uint32_t VectorEG_OnDataRequest(bool hemisphere) {return VectorEG_instance[hemisphere].OnDataRequest();}
-void VectorEG_OnDataReceive(bool hemisphere, uint32_t data) {VectorEG_instance[hemisphere].OnDataReceive(data);}
+uint64_t VectorEG_OnDataRequest(bool hemisphere) {return VectorEG_instance[hemisphere].OnDataRequest();}
+void VectorEG_OnDataReceive(bool hemisphere, uint64_t data) {VectorEG_instance[hemisphere].OnDataReceive(data);}
